整理代码架构 #1
