+++ wooden_judge.h
@@ -1,0 +1,14 @@
+//
+// Created by STUDENT on 2023/3/31.
+//
+
+#include "wooden_test.h"
+#include "wooden_status.h"
+#include "wooden_skill.h"
+
+#ifndef WOODEN_GAME_WOODEN_JUDGE_H
+#define WOODEN_GAME_WOODEN_JUDGE_H
+
+void passon(const tutil::TESTN &test);
+
+#endif //WOODEN_GAME_WOODEN_JUDGE_H
