+++ wooden_start.cpp
@@ -1,0 +1,15 @@
+//
+// Created by STUDENT on 2023/3/31.
+//
+
+#include "wooden_start.h"
+
+//核心初始化
+void wooden_initialization()
+{
+    //检查网络连接
+
+    //
+
+    return;
+}
