+++ wooden_start.h
@@ -1,0 +1,9 @@
+//
+// Created by STUDENT on 2023/3/31.
+//
+
+#ifndef WOODEN_GAME_WOODEN_START_H
+#define WOODEN_GAME_WOODEN_START_H
+void wooden_initialization();
+
+#endif //WOODEN_GAME_WOODEN_START_H
